Update DatafeedService.ts
Increase threshold of pilot difference. It occured today, that the same timestamp was present 46 times. This indicated a Datafeed fault. However, after the issue had been fixed, the difference in pilot numbers was >75, resulting in the datafeed-cache never recovering. 
Increasing the difference to 500 should solve this issue, whilst still having the check of massive discrepancies between updates. Potentially this can be removed in the future.